refactor(svg colors): moved svg colors to constants module
